# Rewrite random projection estimators