--- a/cpp/tests/CMakeLists.txt
+++ b/cpp/tests/CMakeLists.txt
@@ -171,10 +171,6 @@ if(all_algo OR randomforest_algo)
   ConfigureTest(PREFIX SG NAME RF_TEST  sg/rf_test.cu ML_INCLUDE GPUS 1 PERCENT 100)
 endif()
 
-if(all_algo OR randomprojection_algo)
-  ConfigureTest(PREFIX SG NAME RPROJ_TEST  sg/rproj_test.cu ML_INCLUDE)
-endif()
-
 # todo: separate solvers better
 if(all_algo OR solvers_algo)
   ConfigureTest(PREFIX SG NAME CD_TEST  sg/cd_test.cu ML_INCLUDE)