This fully rewrites the `cuml.random_projection` module to make use of cupy instead of C++ from libcuml. The motivations for this were:

- The previous implementation didn't expose the fit `components_` array at the python level, meaning it was impossible to pickle the model. This made it impossible to save a fit model, something data scientists will want to do.
- The previous implementation also was written in a way that made it fail to work with `sklearn.clone`.
- The previous implementation didn't properly support sparse inputs, something the sklearn implementation did.

While all of these could have been fixed while keeping the implementation in C++ (with some major refactoring), there also wasn't a strong reason for the implementation to be done in C++ anyway. The models here are simple and may be implemented straightforwardly using cupy primitives without added performance cost. This reduces our
maintenance burden while also fixing bugs.

Note that this doesn't come with a performance cost. Rather, a quick benchmark shows we're roughly 4x faster for fitting
`GaussianRandomProjection`, and 14x faster for fitting `SparseRandomProjection`. For `transform` we're roughly the same as before. All this with less code and more features.

<details>

<summary>Benchmark Script</summary>

```python
from time import perf_counter
import cuml
import numpy as np


def bench(model, method, *args):
    func = getattr(model, method)
    # Run once to warmup
    func(*args)
    # Run a few times to average
    start = perf_counter()
    for _ in range(3):
        func(*args)
    t = perf_counter() - start
    print(f"{type(model).__name__}.{method}: {t / 3:.2f} s")


X = np.random.default_rng(42).random((10_000, 10_000))
models = [cuml.GaussianRandomProjection(), cuml.SparseRandomProjection()]
methods = ["fit", "transform"]
for model in models:
    for method in methods:
        bench(model, method, X)
```

</details>


**This PR**

```
$ python bench.py
GaussianRandomProjection.fit: 0.02 s
GaussianRandomProjection.transform: 3.09 s
SparseRandomProjection.fit: 0.03 s
SparseRandomProjection.transform: 0.33 s
```

**Previous Implementation**

```
$ python bench.py                                                                                                                        
GaussianRandomProjection.fit: 0.09 s
GaussianRandomProjection.transform: 3.09 s
SparseRandomProjection.fit: 0.42 s
SparseRandomProjection.transform: 0.32 s
```

Fixes #7070
Fixes #6069
Fixes #2390