@jcrist 2025-08-19
I've pushed some updates to expand the test coverage here. I'm now happy with the PR, should be ready for merge IMO.

> would love to have a test on large-ish data (say >60GB) to check we're not missing anything, can help with running a quick couple tests.

I ran a couple local checks on large datasets and everything seems to hold up fine.

```python
import cupy as cp
import cuml

X = cp.random.uniform(size=(10_000_000, 1500), dtype="float32")

def check(cls):
    model = cls(random_state=42, n_components=200)
    model.fit(X)
    model.transform(X)

for cls in [cuml.GaussianRandomProjection, cuml.SparseRandomProjection]:
    check(cls)
```

This runs fine on this PR, but fails _immediately_ on main due to out-of-memory errors (the old implementation makes some unnecessary copies, increasing memory pressure).

Fundamentally random projections are two operations:

- Generation of a random matrix of a size << than the dataset (either dense or sparse)
- A single matrix multiplication (either dense or sparse)

Given a visual inspection of the code here I also wouldn't expect there to be issues (beyond upstream cupy bugs, if any).

> can help with running a quick couple tests.

If you have additional specific things you want to see or test, please do (or let me know _specifically_ what would alleviate your concerns).